[IMP] Improvement in default behaviour of avatax calculation
Currently Avatax calculates tax for SO Lines and Invoice Lines, even when the default placeholder AVATAX
tax is not added to the line, which means lines for Delivery and Down Payment get taxes added even when not needed.
This commit will only compute the avatax if the line has one tax with is_avatax set.